Works with all devices

+++ app/src/main/java/com/zedevstuds/itunesfinder/network/list_of_albums/AlbumListModel.kt
@@ -3,6 +3,6 @@ package com.zedevstuds.itunesfinder.network.list_of_albums
 import com.zedevstuds.itunesfinder.network.list_of_albums.AlbumModel
 
 data class AlbumListModel(
-    val resultCount: Double,
+    val resultCount: Long,
     val results: List<AlbumModel>
 )
